Append authorization and token headers to thumbnail img requests
The headers were added using a service worker.

For creating the service worker:
https://www.twelve21.io/how-to-access-images-securely-with-oauth-2-0/

Adding custom HTTP headers to the service worker script response:
By default, the scope of the service worker (the pages making the requests that are then intercepted by the service worker)
is required to be within the directory containing the service worker script. To set the scope outside of the script's
directory, the 'Service-Worker-Allowed' header needs to be added to the response when the service worker script is requested.
https://drupal.stackexchange.com/questions/307079/how-to-return-service-worker-with-custom-http-header-from-custom-module